feat: feature to take into account contact's time zone when sending mails

+++ src/cron-jobs/jobs/sent-pending-mails.ts
@@ -5,6 +5,7 @@ import ScheduledMailsService from '../services/scheduled-mails.service';
 
 import MailSender from '../../infrustructure/services/mail/mail-sender.service';
 import MailComposer from '../../infrustructure/services/mail/mail-composer.service';
+import MailTimeCoordinator from '../../infrustructure/services/mail/mail-time-coordinator.service';
 
 
 const sentPendingMails = async () => {
@@ -12,15 +13,18 @@ const sentPendingMails = async () => {
 
     //TODO: Handle an error if it was not possible to send a mail
     for (const processedSheduledMailData of pendingMails) {
+        
         const { contactId, id, templateId, ...scheduledMailData } = processedSheduledMailData;
-
         const contactData = await ContactDataService.retrieveContactData(contactId);
-        const composedMail = await MailComposer.composeMail(contactData, templateId);
 
-        await MailSender.sentComposedMail(contactData.email, composedMail);
+        if (MailTimeCoordinator.isTimeToSendMail(processedSheduledMailData, contactData)) {
+            const composedMail = await MailComposer.composeMail(contactData, templateId);
 
-        await ScheduledMailsService.deletePendingMail(id);
-        await SendedMailsService.addSendedMail(processedSheduledMailData);
+            await MailSender.sentComposedMail(contactData.email, composedMail);
+    
+            await ScheduledMailsService.deletePendingMail(id);
+            await SendedMailsService.addSendedMail(processedSheduledMailData);
+        }        
     }
 };
